requirements.txt
# FinAgent - Dependencies
# IIT Bombay Techfest Hackathon

# Core
playwright>=1.40.0

# AI/LLM Integration (New SDK)
google-genai>=1.0.0

# OpenAI (PAID - optional fallback)
openai>=1.6.0

# Web Server for Dashboard
fastapi>=0.109.0
uvicorn>=0.25.0
websockets>=12.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
aiofiles>=23.2.0

# Security
cryptography>=41.0.0

# Scheduling (for recurring transactions)
apscheduler>=3.10.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
black>=23.12.0
